Improve Arabic word splitter
Several fonts (e.g. Arial Unicode MS) contain medial form for Alef, Dal, Thal, Ra', Zay, Waw. In this case we have to verify that it is not the same glyph as for final form. If the medial and final forms are equal, the glyph has no medial form.

DEVSIX-313