Refactor PdfContentStreamProcessor: images handling refactoring, fixed resources handling
Several changes were introduced:
1. Inline images are parsed in PdfContentStreamParser now.

2. Inline images are handled through operator (EI), just in the same way as other objects and actions in content stream are handled

3. Removed PdfImageObject class: ImageRenderInfo now simply return ImageXObject for both inline images and xObject images. Here are the reasons:
PdfImageObject was a simple data-container class with some logic to convert pdf images to native ones. Converting images to native is useful, but it probably should be implemented in some other place, not just for content parsing module; it will probably take PdfImageXObject as argument. From other perspectives, ImageXObject contain the same functionality and also is convenient for user (i.e. it is simple to copy image to other document), also it fits in existing iText6 image hierarchy. Concerning inline images, they would be wrapped in PdfStream, which in its turn would be wrapped in PdfImageXObject. Technically it would look like just as usual ImageXObject, but user can define if it was inline image by specific field in ImageRenderInfo.

4. Removed ResourceDictionary class: resources inheritance was interpreted in incorrect way. According to spec if resource dictionary is absent in the page's dictionary or in the formXObject dictionary, then it is inherited from object's parents. But when the nearest resource dictionary is found, this dictionary is interpreted as current resource dictionary. The keys shouldn't be searched for in other dictionaries.