Separate MCR objects and canvas tag objects. Support canvas tag property lists and structElements attributtes
From now, PdfTagStructure provides not actually the CanvasTag instances, but rather an intermediate PdfTagReference object. This object could be passed to the PdfCanvas in order to begin marked content sequence. The reason to add this intermediate object is to avoid the possible scenario, when user creates CanvasTag using PdfTagStructure(this new canvas tag will have a mcid assigned to it) and then adds this CanvasTag to canvas several times, which will produce several distinct marked content sequences with same mcid. When PdfTagReference is used, actual CanvasTag is created only in PdfCanvas#openTag method.