Layout: optimize handling of renderers' parents
Parents used to be assigned right on adding a child elements to a renderer which could lead to memory leaks in some cases as well as inconsistencies on splitting. After splitting, parents of overflow renderers used to be manually reassigned in some cases, which caused additional overhead. Now parents must be manually set before any operations that require parent's knowledge, such as property retrieval, or any operations that imply property retrieval