Restored backward compatibility.
The only issue that was ignored was the one with changing the type of protected field for TaggedPdfReaderTool.java. This was done in the sake of automated porting. The similar issue with XmlDomWriter.java was fixed because whole this package is on manual porting. Also were added @deprecated annotations for fields and methods that are supposed to be removed at some point.

DEVSIX- 640