Fix parsing of Pattern colors from content
The failure was in case current colorspace is set to /Pattern with parameter, e.g. [/Pattern /DeviceRGB], but the pattern being set is itself colored, i.e. no color components should be specified. The color components used to be parsed in a forced way, now they are parsed only in case the pattern itself is uncolored