Make view instances of RandomAccessFileOrArray be thread-safe between each other and also with the original instance
The createView() method is used by us in order to create an independent instance which could be used in multithreaded context. This same use case of this method I've also encountered in some stackoverflow question (http://stackoverflow.com/questions/33235686/multi-threaded-approach-to-convert-tiff-to-pdf-using-itext). However, view instances share their byte sources, which might be not thread-safe. So, in order to fix this, their shared byte source is put in thread-safe wrapper on view creation.

DEVSIX-996