Reverse order of Style lookup in AbstractElement.getProperty.
Look in styles for a property in the reverse order the Styles are added. This enables a client to
add a new Style to an Element that can override the effect of previously added Styles. Otherwise
there is no way to change the property using a Style without changing the underlying Style.  The
lookup order doesn't seem to be documented anywhere, so while this is an incompatible change, it
doesn't seem to affect the documentation.
The existing tests of addStyle in StylesTest are unaffected, but the new test would have failed with
the old code.