Allow LIST_SYMBOL to be overrriden by Styles.
As currently implemented, it is not possible to change LIST_SYMBOL using a Style. Since all List
constructors set the LIST_SYMBOL property directly on the List, this property setting has precedence
over any Style setting.

This change makes LIST_SYMBOL a default property so that the List's Style properties have
precendence over the *default* LIST_SYMBOL. If a symbol is specified in the List constructor, it
still has precedence.  See ListTest.listSetSymbol.

AbstractRenderer.getProperty first calls modelElement.getProperty (which now returns null in this
case) and then calls modelElement.getDefaultProperty, which returns Text("- ").
However, this change is not backwards-compatible since it is possible that a client will call
List.getProperty and get null instead of Text("- ").

Two alternatives:
- new List(null) could call deleteOwnProperty(Property.LIST_SYMBOL)
- leave things alone and require the client to call deleteOwnProperty