Fix CompareTool on Mac and handling of spaces in names
The original Compare parameter string included double-quotes
which were ignored on Windows but caused compare to always
fail on Mac.

Correctly handle files that include spaces by adding single quotes
around filenames in both the gs and compare parameter strings
and then modifying SystemUtil.runProcessAndWait to ignore spaces
within those quotes when tokenizing the parameters.
compareToolErrorReportTest04 tests processing of a filename that
includes a space.

Reverse check of result of running compare in compareImagesOfPdfs:

Compare returns 0 when there is no change, and
SystemUtil.runProcessAndWait returns true when the result is
zero, so the check of the return value should refer to the
differences output when the the return value is false.  You can
see this in the existing compareToolErrorReportTest01 where
running convert on the first page results in differences but the
"Please, examine" message didn't show up before (and does now).

Change failures when executing gs or compare to be exceptions. By
clearly distinguishing between these failures and detecting
differences, it's easier to detect problems like spaces in
filenames.
(I wanted to further distinguish by the result of running compare, but
compare returns 2 whenever the comparison fails, which includes when
the images are of different dimensions. This means that other
compare failures (such as having an extra " in the filename)
aren't easily detected since compare returns 2 in all these cases.)