Improve rotation handling
Now rotation angle is taken into account in heuristic used in min-
max-width calculations for block elements and paragraphs. Calculated
width is checked by actual layout. In worst case (element width
isn't set, its rotation angle is not close to 90 deg) there can be
two more layouts during min-max-width calculation.

During layout of paragraph we now try to set proper width of layout
area so that after rotation it will fit in initial width. This
calculations are using the same heuristic as in min-max-width
calculations. If the width wasn't set on element we will need to get
min-max-width of not rotated element and call one layout, for this
heuristic to work.

Also now width of rotated elements won't be chopped to parent width,
since it can be way bigger, for elements with rotation angle close
to 90 deg. Also now rotated paragraph will be transferred to the
next page only in case it doesn't fit by height.

DEVSIX-1241