Partial fixes for floats and max_width %
This is a work in progress. It has some fixes for dealing with the
interaction of floats and max_width, but there is still a problem
with max_width for Paragraph, because when the Paragraph is not a
float, its width is always the width of its parent (or its parent's
max_width, if that is specified).
Also, this has a "fix" for when an image's max_width is 100%. layout()
returns a width of the content, not including the border, so the
width is not wide enough to draw the image and border when the image
width or max_width is 100%. (I only fixed for max_width, but both
should work the same way. One possibility is to special-case 100%;
another is to add a property BOX_SIZING like the CSS property. Then
if the value is border-box (rather than the default content-box),
the right thing would happen.