Add convenient methods for setting annotation default appearance
The new class provides setters for font-color, font-size and font
itself.
Note that only standard font names that don't require font resources
are supported.
Note that it is possible to create annotation with custom font name
in DA, but this require manual resource modifications (you have to
put font in DR of AcroForm and use it's resource name in DA) and
only Acrobat supports that workflow.

DEVSIX-1401