Make checkbox field take into account user-specified font-size; auto-size font by default; fix bugs
Also rewrite the approximation logic for font-size autoscaling: paddings calculation
cannot result in negative values for font-size now, additionally padding calculations
are now not dependent on borders width.

Remove limitation on font-size writing in PdfCanvas. This check originated from
iText 5, where it was stated that some viewers might crash because of the too
small font-size. However it's convenient to write /DA key for acro fields
via PdfCanvas and in this content zero font-size is absolutely valid value.
In general it seems that it's should be user responsibility to care of such things.

Fix checkbox field toggling. Previously when it's state has been set to Off, checkbox
appearance for "on" state was lost.

DEVSIX-1861