Add event processing API
 * The EventDataHandler is a class intended for some heavy
concurrent event operations (like writing to database or file).
 * The IEventDataFactory interface is intended for adding additional
information to the event (like system info or timestamp) and also
for performing event filtering.
 * The IEventDataCache interface is intended for event cashing,
for example it can merge the EventData with same signature by
summing count.
 * It also determines the order of EventData processing. For example
it can retrive data based on the time of there signature
registration, or it can retrive data with the biggest count first.

DEVSIX-1958