Remove limitation on font-size writing in PdfCanvas
This check originated from iText 5, where it was statedthat some viewers
might crash because of the too small font-size. However it's convenient
to write /DA key for acro fields via PdfCanvas and in this content
zero font-size is absolutely valid value. In general it seems that
it's should be user responsibility to care of such things.

DEVSIX-1916
DEVSIX-1861