Fix OCSPVerifier responder certificate verification
The issuer cert passed to the OCSPVerifer is considered valid and trusted, it's not OCSPVerifier responsibility to check the complete certificate chain. It only verifies if given certificate is not revoked according to the OCSP service and ensures that OCSP response is genuine. Thus, only authorized responder certificate (if it is used) is checked for validity and possible revocation against signature date.

Also fix CRL entry validity check in CRLVerifier: CRLs issued only recently can still be used for validating certificates used at any time in the past.

Additionally introduce initial implementation of new signature tests utility for creating test certificates programmatically on the fly.

DEVSIX-2710