Allow overriding test timeout for classes extending from ITextTest
Inheritance is not taken into account for Junit @Timeout rules.
This means that if several @Timeout rules are present in the class
inheritance chain, and even if timeout is specified as an attribute
of the @Test annotation, the minimal timeout would be decisive.
This fix allows classes extending from ITextTest setting bigger
timeout for tests. Previously it was only possible to make timeout
smaller due to the fact described above.
See also https://github.com/junit-team/junit4/issues/1126