Use initial bytes to check if byte[] is WmfImage
When creating a WmfImage from a byte array use the first few bytes
(instead of a null URL) to check whether it is a WMF image.

Loading a WmfImage from a filepath or URL compares the first few
bytes from the file. For a byte array the same method was used,
which failed because no URL is provided.