Make PdfImageXObject more independent of underlying object flushing
Previously PdfImageXObject reusing after flushing relied on wrapper flush() being called, so width and height were preserved. Now they are extracted on wrapper creation. Image XObject dictionary width and height are based on the underlying image bytes and are not expected to be changed: we are not aware of such scenarios and if someone manually changes those entries, they would need to create a new wrapper for changed object.

Also, PDF image decoding functionality is extracted to a separate class.

DEVSIX-2835