Do not trim leading spaces in single line and multiline text fields
Previously, leading spaces were trimmed because this is the default
layout behavior. However, Acrobat does not trim leading spaces and
thus there was a difference between the appearance generated by iText
and the one that Acrobat generates when you enter the field. We now
follow Acrobat and do not trim leading spaces

DEVSIX-3539