Revert to typo_ascender scale coefficient usage for Type1Fonts for HTML mode
In standard PDF fonts according to iText metrics typo ascender and
typo descender don't fully include some of the font glyphs.
Additionally, in HTML mode for AD < 1150 in case of 'line-height: normal'
the line is expanded vertically a bit and according to CSS specification,
this expansion doesn't preserve ascent/descent aspect ratio. So for standard
PDF fonts (for which AD < 1150) the baseline appeared closer to the line center
resulting in not optimal appearance. This issue is still present for
other fonts, but at least not that significant for the default one.

Preserving the aspect ratio even only for normal mode would result
in inconsitency with CSS specification and our own behavior.
Artificially increasing ascent/descent up to our normal line height (1.15)
for all fonts will also lead to unexpected rendering.

DEVSIX-3661