Improve the security of XML parsers
Improve the security of XML parsers to prevent XML attacks like XML bombs and XXE attacks.
Set default configuration of xml parsers to throw an exception when it find DTD in the XML and implement factory logic to give an opportunity for customer to override the way of creation xml parsers.

DEVSIX-3270