Revert increased strictness for getObjectBoundingBox
Now in case of undefinied bbox a null can be returned. This was the original implementation and this
allows to handle both valid and invalid cases in general fashion. E.g. for gradient color logic, we can
be simply more lenient towards invalid inputs and just ignore meaningless gradient definiitions.

DEVSIX-5427