Introduce setNeutralRole() convenience API
This method was added to a number of common layout object classes to
replace `.getAccessibilityProperties().setRole(null)` with a chainable
`.setNeutralRole()` call, which reduces boilerplate for some complex
tagging workflows.

Additionally, this commit adds documentation to `setRole()` to explain
the effect of marking the role as `null`.

Finally, internal uses of `.setRole(null)` have been replaced with calls
to this new method, and tests added accordingly.

RES-630 DEVSIX-6402