Fix bug/issue when subtype is null on annotations
There are some PDF files where the getSubtype function returns null, which doesn't implement equals() and that leads to an error on line 120. This is something it happens to my team in our software using IText, checking if the sub type of the annotation is null prevent from an error and it just doesn't copy the field.