Clean up naming in signing internals
Many fields and methods in PdfPKCS7 have names that are misleading,
wrong or no longer make sense in a setting where RSA is not the only
game in town.

This commit renames a number of internal fields to be less confusing
and updates the documentation accordingly. For compatibility reasons,
the public API has not been modified.