Ensure compatibility with BC-FIPS on JDK 8
There's an issue causing certificate loading to fail when BC-FIPS is
used with a CertificateFactory from the default security provider on
JDK 8, when the subject key is an EdDSA key.

This commit changes that by ensuring that the currently relevant BC
provider is always used when constructing CertificateFactory objects in
the 'sign' module. The change was applied to both BC and BCFIPS for
consistency.