Fix width calculation for TextRendererInfo by changing PdfSimpleFont logic
Attempt to get glyph by code recorded in font-program.
We used to first get glyph based on unicode, because glyphs can be identified by unicode
in font program (even though not uniquely). Let's go the other way around, let's first
try to get the glyph by code in a hope that font program was created in accordance to the give PdfFont
encoding. If unicode matches - then guess quality is more or less the same as if we were going the
other way around.

DEVSIX-4784